backend/requirements.txt
fastapi
uvicorn
google-cloud-firestore
sentence-transformers
vaderSentiment
feedparser
requests
python-multipart
